+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 0.10.12
+* Added features to view authors of a file or current selection.
 * Updated German translation by [Marcel Gerber](https://github.com/SAPlayer)
 * Push dialog shows masked password.
 * Tabs are now properly displayed in the diffs respecting Brackets "tabSize" preference.